+++ custom_components/xplora_watch/entity.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 
 import logging
 
-from datetime import datetime
 from typing import Any
 
 from .helper import XploraUpdateTime